requirements.txt
djangorestframework
django
mysqlclient